remove util.setWebGLType
The problem is the code was designed so there's just "webgl" so it calls util.getWebGLContext and always created a "webgl" context that's not compatible with "webgl2".

The quick and dirty solution was to set global which kind of context to make but that fails on pages that use both webgl and webgl2 contexts.